コードレビューの指摘された内容の修正
- Homeコンポーネント別のファイルに切り出し
- スレッド入力フォームがからの状態のときはボタンを押せないように<button>にdisabledを追加
- useNavigateのフックを利用してスレッド新規作成後にトップページに遷移できるようにした
- Headerコンポーネントは{children}が構造上<header>内の要素であると明示的にするために<header>内に移動
