+++ src/App.tsx
@@ -1,45 +1,8 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { useEffect, useState } from 'react'
-import { Link } from 'react-router-dom';
-import './App.css'
-import Header from './Header/Header';
+import Home from './Home/Home';
 import ThreadsCreate from './ThreadsCreate/ThreadsCreate';
 
-type Threads = {
-  id: number;
-  title: string;
-}
-
-const Home = () => {
-  const [threads, setThreads] = useState<Threads[]>([]);
-
-  useEffect(() => {
-    fetch("https://railway.bulletinboard.techtrain.dev/threads?offset=20")
-      .then(res => res.json())
-      .then(data => setThreads(data))
-      .catch(error => {
-        console.log("スレッドのデータを取得できません", error)
-      })
-  }, []);
-  console.log(threads)
-  return (
-    <>
-      <Header>
-        <Link to="/threads/new" className='threadCreateLink' >スレッドをたてる</Link>
-      </Header >
-      <section className='threadContainer'>
-        <h1>新着スレッド</h1>
-        <ul>
-          {threads.map(thread => (
-            <li className='threadCord' key={thread.id}>{thread.title}</li>
-          ))}
-        </ul>
-      </section>
-    </>
-  )
-}
-
-function App() {
+export default function App() {
   return (
     <Router>
       <Routes>
@@ -49,5 +12,3 @@ function App() {
     </Router>
   )
 }
-
-export default App
