+++ src/Home/Home.css
@@ -1,4 +1,5 @@
 body {
+  margin: 0;
 }
 
 .headerContainer {
