+++ src/ThreadsCreate/ThreadsCreate.tsx
@@ -1,14 +1,14 @@
 import { useState } from "react";
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import "./threadsCreate.css";
 import Header from "../Header/Header";
 
 export default function ThreadsCreate() {
   const [createTitle, setCreateTitle] = useState("");
-  // TODO:response後のメッセージ表示するStateを作成
+  const navigate = useNavigate();
+
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-
     const postData = {
       title: createTitle
     }
@@ -20,14 +20,19 @@ export default function ThreadsCreate() {
           'Content-Type': 'application/json',
         },
         body: JSON.stringify(postData),
-      });
-      console.log("スレッドの作成完了:", postData)
+      })
+        .then(res => res.json())
+        .then(data => {
+          alert("スレッドの作成完了:" + data.title);
+          setCreateTitle("");
+          navigate("/");
+        })
     }
     catch (error) {
-      console.log("スレッドのデータを取得できません", error)
+      console.log("スレッドが作成できませんでした。", error)
     }
-
   }
+
   return (
     <>
       <Header>
@@ -35,17 +40,19 @@ export default function ThreadsCreate() {
       </Header>
       <section>
         <h1>スレッド新規作成</h1>
-        <form onSubmit={handleSubmit}>
-          <label>
-            スレッドタイトル
-            <input type="text"
-              value={createTitle}
-              name="threadTitle"
-              onChange={(e) => setCreateTitle(e.target.value)}
-              placeholder="タイトルを入力してください" />
-          </label>
-          <button type="submit">作成</button>
-        </form>
+        <label>
+          スレッドタイトル
+          <input type="text"
+            value={createTitle}
+            name="threadTitle"
+            onChange={(e) => setCreateTitle(e.target.value)}
+            placeholder="タイトルを入力してください"
+          />
+        </label>
+        <button
+          onClick={handleSubmit}
+          disabled={createTitle === ""}
+        >作成</button>
       </section>
     </>
   )
